+++ test/_env/app/xml/webapp/Component.js
@@ -1,0 +1,7 @@
+sap.ui.define(["sap/ui/generic/app/AppComponent"], (AppComponent) =>
+  AppComponent.extend("com.sap.cds.app.xml.Component", {
+    metadata: {
+      manifest: "json",
+    },
+  })
+);