+++ test/batch-test.js
@@ -49,9 +49,9 @@ describe("batch", () => {
     expect(responses.length).toEqual(3);
     expect(responses.filter((response) => response.statusCode === 200).length).toEqual(3);
     const [first, second, third] = responses;
-    expect(first.body.d.results.length).toEqual(6);
+    expect(first.body.d.results.length).toEqual(7);
     expect(first.contentTransferEncoding).toEqual("binary");
-    expect(second.body.d.results.length).toEqual(6);
+    expect(second.body.d.results.length).toEqual(7);
     expect(second.contentTransferEncoding).toEqual("binary");
     expect(third.body.d.hasOwnProperty("results")).toEqual(false);
     expect(third.body.d.ID).toEqual(ID);
@@ -90,7 +90,7 @@ describe("batch", () => {
     expect(responses.length).toEqual(3);
     const [first] = responses;
     expect(first.body.d.results).toBeDefined();
-    expect(first.body.d.results.length).toEqual(8);
+    expect(first.body.d.results.length).toEqual(9);
     expect(first.body.d.results[0]).toBeDefined();
     expect(first.body.d.results[0].name).toBeDefined();
     expect(first.body.d.results[0].__metadata).toBeDefined();